+++ src/main/php/xp/unittest/sources/ClassSource.class.php
@@ -1,6 +1,7 @@
 <?php namespace xp\unittest\sources;
 
 use lang\XPClass;
+use unittest\{TestCase, TestMethod};
 
 /**
  * Source that load tests from a class filename
@@ -29,11 +30,13 @@ class ClassSource extends AbstractSource {
   public function provideTo($suite, $arguments) {
     if (null === $this->method) {
       return $suite->addTestClass($this->testClass, $arguments);
-    } else {
+    } else if ($this->testClass->isSubclassOf(TestCase::class)) {
       $suite->addTest($this->testClass->getConstructor()->newInstance(array_merge(
         [$this->method],
         (array)$arguments
       )));
+    } else {
+      $suite->addTest(new TestMethod($this->testClass, $this->method, (array)$arguments));
     }
   }